Remove length integrity check in funzip.
funzip hasn't been touched in 20 years, and does not support the
Zip64 zip format extensions. This commit bypasses the uncompressed
length check, where the length can be stored in a Zip64 extended
information extra field as eight bytes, instead of the four-byte
field in the local header. Such entries can now be successfully
decompressed. The CRC integrity check is still performed.

This is a temporary application of duct tape for this issue. The
code will be updated to handle the modern formats for large files,
as well as handle streaming data descriptors.