A simple patch to detect zip bombs.
This works by limiting the total amount of compressed bytes that can be
consumed in a single run before we consider the file a zip bomb to the
total size of the zip file (times two).

Unlike the cover-based solution, this doesn't detect every case where file
entries overlap, but importantly, it does detect cases where quadratic runtime
is caused by overlapping file entries so they are processed more than twice.

Because this fix allows each part of the zip file to be covered twice,
it is about twice as slow at detecting zip bombs. This is is confirmed by
by benchmarking with the zip bomb files from
https://www.bamsoftware.com/hacks/zipbomb/

              cover       simple
    zbsm.zip  0m00.159s   0m00.244s
    zblg.zip  0m20.460s   0m40.735s
    zbxl.zip  1m52.988s   3m46.859s

So while this patch is slower to detect actual zip bombs, it has the
benefits of being extremely simple, adding no overhead whatsoever, and
not being exploitable in any way.

In conclusion, it's a simple yet effective way to mitigate CVE-2019-13232.