Bugfix: set umask(022) when creating output files
This was clearly intended to be enabled on Linux/UNIX but because
the ATH_BEO_UNX variable was misspelled as ATH_BE_UNX it never was.

Setting umask limits the permission bits of the output file to 0644,
preventing other users from modifying the file while it's being
written.

This does not affect the final permission bits, since these are set
with (f)chmod() in closefile() defined in unix/unix.c.

The security benefit is likely to be small in practice, since on
most systems, umask = 022 by default anyway.