Remove unused code that was used by obsolete operting systems and compilers.
This is a massive change that deletes around 5000 lines (about 1/6th) of the
source code, but without affecting the binaries that get compiled on Linux.

Unfortunately, there is a bit of churn due to preprocessor directives
being re-indented even in places where it wasn't strictly necessary, but I'm
okay with it because the result is a lot clearer. View the commit diff without
whitespace changes if you find it distracting.