Always enable ACORN_FTYPE_NFS.
This adds the -F option to unzip, which preserves Acorn file type information.
It's not really a common feature, but most distros already enabled this, and
there is not a particularly good reason to leave it out. Note that this doesn't
change default behavior at all; it only adds an extra command line flag.