Remove _MBCS code (support for multi-byte strings).
This was only used by the removed Human68k, Win32 and Cygwin ports.

Modern systems are usually UTF-8, which is technically a multibyte
encoding, but it generally supports handling strings as byte strings.
In any case, the UNIX/Linux ports didn't use the _MBCS flavor
regardless of what character encoding is used.