Enable NOMEMCPY by default.
This fixes bugs in explode.c and inflate.c. Note that most distros
(including Debian, Fedora and Arch Linux) already defined NOMEMCPY
so this doesn't change anything there.

tl;dr: the use of memcpy() here violated the requirement that
the source and destination regions do not overlap.

To go in a bit more detail why this change is necessary, let's
take a closer look at the removed code:

    if (w - d >= e)       /* (this test assumes unsigned comparison) */
    {
      memcpy(redirSlide + w, redirSlide + d, e);
      w += e;
      d += e;
    }

Here, d and w are the source and destination offsets for the
copy, and e is the number of bytes to copy. (e has been calculated
so the copy won't exceed the buffer size) When d <= w, the if-
condition correctly verifies that the source and destination
regions do not overlap. However, since this is a circular
buffer, it's also possible that d > w, and in that case, the
regions do overlap.

This is less likely to cause problems in practice because
typically memcpy() works from left to right, which causes problems
when the source pointer precedes the destination pointer, but
not the other way around. This probably explains why this bug
was not detected sooner. However, there is no guarantee that
memcpy() works like this, and some implementations might not,
or might check if their arguments are valid and abort() or not.

In short: this code was wrong and should not be used.

It would be possible to fix the condition with:

    if (w - d >= e && d - w >= e) { ..  }

(Note that this calculation is correct only because w, d and e are
unsigned integers, otherwise integer overflow would cause undefined
behavior.)

However, I don't think it's worth optimizing this. In the interest
of stability, I'd rather commit to the behavior that already existed
in practice, which is well-tested by now.

Note that OpenBSD fixed this bug by replacing memcpy() with
memmove(): https://codeberg.org/OpenBSD/ports/commit/89a66a1

This is also correct, but since the else-clause is effectively
an adhoc memmove() implementation already, it seems simpler to
just delete the first clause. Alternatively, the entire section
could be replaced by a single call to memmove(), but again,
I want to avoid making unnecessary changes.