Clean up and fix bugs in EF_IZUNIX3 parsing.
The old code had two bugs:

  - There were size checks missing wich could lead to reading out of bounds
    if the uid_size field had a high value (e.g. 255). I wasn't able to make
    unzip crash by abusing this, but it still seems prudent to fix it.

  - read_ux3_varint() would incorrectly return TRUE for unsupported sizes,
    while leaving the return value uninitialized. This would probably lead
    to random values being used in the call to fchown() later.

(I plan to enable IZ_HAVE_UXUIDGID by default later, hence these fixes.)

I would like to add a test for this change but it's too difficult because
it requires either root access or a specific setup where the user/group
the tests run under is known in advance. I decided it's not worth the
trouble.