Enable IZ_HAVE_UXUIDGID by default.
This enables `unzip -X` to restore UID/GID information (which requires either
root access, or else it can only restore uid/gids where the uid is the current
user, and the gid is a group that user is a member of).

Debian and Fedora already had this enabled manually. Arch Linux did not, and
as a result unzip did not behave as documented (i.e., unzip -X didn't restore
ownership as it should).