zipgrep: Replace `echo` with `printf "%s\n"` for increased portability.
In Bash, "echo -n" doesn't print anything, because bash interprets "-n" as
an option, and there is no way to disable this. Other shells don't have the
same options, so there is no general way to avoid this issue.

This issue was reported here:
https://gitlab.archlinux.org/archlinux/packaging/packages/unzip/-/issues/1

The recommendation is to use printf instead:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/echo.html#tag_20_37_16

printf avoids this issue when the format string is statically specified, which
makes it easy to ensure it doesn't start with a '-', and any following arguments
are not interpreted as option.